+++ src/main/java/com/sbproject/piclocket/repository/PhotoRepository.java
@@ -1,9 +1,13 @@
 package com.sbproject.piclocket.repository;
 
 import com.sbproject.piclocket.model.Photo;
+import com.sbproject.piclocket.model.PhotoStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.UUID;
 
 public interface PhotoRepository extends JpaRepository<Photo, UUID> {
+
+    List<Photo> findByStatus(PhotoStatus status);
 }